Velocidade Player

+++ src/pong/Ball.java
@@ -65,7 +65,7 @@ public class Ball {
 		
 		Rectangle bounds = new Rectangle((int)(x+(dx*speed)), (int)(y+(dy*speed)), width, height);
 		
-		Rectangle boundsPlayer = new Rectangle(Game.player.x, Game.player.y, Game.player.width, Game.player.height);
+		Rectangle boundsPlayer = new Rectangle((int)Game.player.x, (int)Game.player.y, Game.player.width, Game.player.height);
 		Rectangle boundsEnemy = new Rectangle((int)Game.enemy.x, (int)Game.enemy.y, Game.enemy.width, Game.enemy.height);
 		
 		if(bounds.intersects(boundsPlayer)) {
